feat(kit-demo): split Feed + Scroll into own tabs, fix Gestures icon, add test plan
